refactor: split joins out
